del allusers and secret resource, refactor filenames, fixed imports, moved models to models folder, removed dotenv from run.py, changed err msg
