+++ .gitignore
@@ -30,6 +30,5 @@ instance/
 .vscode
 
 
-#TODO add .env to your gitignore
 server/.env
