+++ server/models/user_model.py
@@ -1,4 +1,4 @@
-from run import db
+from server.run import db
 from passlib.hash import pbkdf2_sha256 as sha256
 
 
@@ -52,20 +52,3 @@ class UserModel(db.Model):
     def verify_hash(password, hash):
         # check given password
         return sha256.verify(password, hash)
-
-
-class RevokedTokenModel(db.Model):
-    # store primary key id and jti(unique identifier of the token). add() method adds token
-    # to the database. is_jti_blacklisted() checks if the token is revoked
-    __tablename__ = 'revoked_tokens'
-    id = db.Column(db.Integer, primary_key=True)
-    jti = db.Column(db.String(120))
-
-    def add(self):
-        db.session.add(self)
-        db.session.commit()
-
-    @classmethod
-    def is_jti_blacklisted(cls, jti):
-        query = cls.query.filter_by(jti=jti).first()
-        return bool(query)
