+++ server/resource/user_logout_access.py
@@ -1,6 +1,6 @@
-from models import UserModel, RevokedTokenModel
 from flask_restful import Resource
 from flask_jwt_extended import (jwt_required, get_raw_jwt)
+from server.models.revoked_token_model import RevokedTokenModel
 
 
 class UserLogoutAccess(Resource):
