+++ server/resource/user_registration.py
@@ -1,10 +1,8 @@
 from flask_restful import Resource, reqparse
 from flask_jwt_extended import (
     create_access_token, jwt_required, get_jwt_identity)
-import sys
-sys.path.append("..")
-from models import UserModel
 import datetime
+from server.models.user_model import UserModel
 
 
 class UserRegistration(Resource):
